# All gather and Reduce Scatter dispatch time optimization 