--- a/ttnn/cpp/ttnn/operations/ccl/all_gather/device/kernels/dataflow/worker_interleaved_ring_gather_receive_writer.cpp
+++ b/ttnn/cpp/ttnn/operations/ccl/all_gather/device/kernels/dataflow/worker_interleaved_ring_gather_receive_writer.cpp
@@ -14,6 +14,29 @@ void kernel_main() {
     // Different per worker receiver writer
     const uint32_t worker_sender_reader_noc_x = get_arg_val<uint32_t>(arg_idx++);
     const uint32_t worker_sender_reader_noc_y = get_arg_val<uint32_t>(arg_idx++);
+    const bool receiver_enabled = get_arg_val<uint32_t>(arg_idx++) == 1;
+    if (!receiver_enabled) {
+        return;
+    }
+    const uint32_t num_transfers = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t num_full_chunks = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t num_pages = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t rem_num_pages = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t output_start_page_idx = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t output_start_addr_offset = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t row_start_idx = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t col_start_idx = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t row_offset = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t col_offset = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t num_rows = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t num_cols = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t last_output_page_offset = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t output_page_offset = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t last_output_addr_offset = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t output_addr_offset = get_arg_val<uint32_t>(arg_idx++);
+    const uint32_t input_start_ring_idx = get_arg_val<uint32_t>(arg_idx++);
+    const bool is_clockwise_direction = get_arg_val<uint32_t>(arg_idx++) == 1;
+
 
     #ifdef SHARDED_MEM_LAYOUT
     uint32_t output_shard_grid_nrows = get_arg_val<uint32_t>(arg_idx++);
@@ -26,42 +49,24 @@ void kernel_main() {
 
     // Compile time Args
     constexpr bool dst_is_dram = get_compile_time_arg_val(0) == 1;
-    constexpr uint32_t num_transfers = get_compile_time_arg_val(1);
-    constexpr uint32_t num_full_chunks = get_compile_time_arg_val(2);
-    constexpr uint32_t page_size = get_compile_time_arg_val(3);
-    constexpr uint32_t output_page_size = get_compile_time_arg_val(4);
-    constexpr uint32_t num_pages = get_compile_time_arg_val(5);
-    constexpr uint32_t rem_num_pages = get_compile_time_arg_val(6);
-    constexpr uint32_t output_start_page_idx = get_compile_time_arg_val(7);
-    constexpr uint32_t output_start_addr_offset = get_compile_time_arg_val(8);
-    constexpr uint32_t row_start_idx = get_compile_time_arg_val(9);
-    constexpr uint32_t col_start_idx = get_compile_time_arg_val(10);
-    constexpr uint32_t row_offset = get_compile_time_arg_val(11);
-    constexpr uint32_t col_offset = get_compile_time_arg_val(12);
-    constexpr uint32_t num_rows = get_compile_time_arg_val(13);
-    constexpr uint32_t num_cols = get_compile_time_arg_val(14);
-    constexpr uint32_t last_output_page_offset = get_compile_time_arg_val(15);
-    constexpr uint32_t output_page_offset = get_compile_time_arg_val(16);
-    constexpr uint32_t last_output_addr_offset = get_compile_time_arg_val(17);
-    constexpr uint32_t output_addr_offset = get_compile_time_arg_val(18);
-    constexpr uint32_t input_start_ring_idx = get_compile_time_arg_val(19);
+    constexpr uint32_t page_size = get_compile_time_arg_val(1);
+    constexpr uint32_t output_page_size = get_compile_time_arg_val(2);
     // Same per worker receiver writer
-    uint32_t sem_addr = get_semaphore(get_compile_time_arg_val(20));
-    constexpr bool is_clockwise_direction = get_compile_time_arg_val(21) == 1;
-    constexpr uint32_t half_cb_n_pages = get_compile_time_arg_val(22);
-    constexpr uint32_t ring_size = get_compile_time_arg_val(23);
-    constexpr bool fuse_op = get_compile_time_arg_val(24);
-    static_assert(half_cb_n_pages > rem_num_pages, "half_cb_n_pages must be greater than or equal to rem_num_pages");
+    uint32_t sem_addr = get_semaphore(get_compile_time_arg_val(3));
+    constexpr uint32_t half_cb_n_pages = get_compile_time_arg_val(4);
+    constexpr uint32_t ring_size = get_compile_time_arg_val(5);
+    constexpr bool fuse_op = get_compile_time_arg_val(6);
+    // static_assert(half_cb_n_pages > rem_num_pages, "half_cb_n_pages must be greater than or equal to rem_num_pages");
 
     #ifdef SHARDED_MEM_LAYOUT
-    constexpr tt::tt_metal::TensorMemoryLayout output_tensor_memory_layout = static_cast<tt::tt_metal::TensorMemoryLayout>(get_compile_time_arg_val(25));
-    constexpr uint32_t output_tensor_shard_grid_height = get_compile_time_arg_val(26);
-    constexpr uint32_t output_tensor_shard_grid_width = get_compile_time_arg_val(27);
-    constexpr uint32_t output_tensor_shard_grid_start_y_logical = get_compile_time_arg_val(28);
-    constexpr uint32_t output_tensor_shard_grid_start_x_logical = get_compile_time_arg_val(29);
-    constexpr uint32_t output_tensor_shard_pages_per_shard_y = get_compile_time_arg_val(30);
-    constexpr uint32_t output_tensor_shard_pages_per_shard_x = get_compile_time_arg_val(31);
-    constexpr bool output_tensor_shard_grid_transposed = get_compile_time_arg_val(32) != 0;
+    constexpr tt::tt_metal::TensorMemoryLayout output_tensor_memory_layout = static_cast<tt::tt_metal::TensorMemoryLayout>(get_compile_time_arg_val(7));
+    constexpr uint32_t output_tensor_shard_grid_height = get_compile_time_arg_val(8);
+    constexpr uint32_t output_tensor_shard_grid_width = get_compile_time_arg_val(9);
+    constexpr uint32_t output_tensor_shard_grid_start_y_logical = get_compile_time_arg_val(10);
+    constexpr uint32_t output_tensor_shard_grid_start_x_logical = get_compile_time_arg_val(11);
+    constexpr uint32_t output_tensor_shard_pages_per_shard_y = get_compile_time_arg_val(12);
+    constexpr uint32_t output_tensor_shard_pages_per_shard_x = get_compile_time_arg_val(13);
+    constexpr bool output_tensor_shard_grid_transposed = get_compile_time_arg_val(14) != 0;
     #endif
 
     constexpr uint32_t cb_id_in0 = tt::CB::c_in0;
@@ -135,7 +140,7 @@ void kernel_main() {
 
 
     for (uint32_t i = 0; i < num_transfers; ++i) {
-        if constexpr (num_full_chunks > 0) {
+        if (num_full_chunks > 0) {
             for (uint32_t c = 0; c < num_full_chunks; ++c) {
 
                 #ifdef SHARDED_MEM_LAYOUT
@@ -145,7 +150,7 @@ void kernel_main() {
                 noc_semaphore_inc(worker_send_reader_semaphore_noc_addr, 1);
             }
         }
-        if constexpr (rem_num_pages > 0) {
+        if (rem_num_pages > 0) {
             #ifdef SHARDED_MEM_LAYOUT
             ASSERT(output_page_idx < output_tensor_shard_pages_per_shard_y * output_tensor_shard_pages_per_shard_x * output_tensor_shard_grid_height * output_tensor_shard_grid_width);
             #endif
@@ -159,36 +164,36 @@ void kernel_main() {
         if (is_clockwise_direction) {
             if (input_ring_idx == 0) {
                 input_ring_idx = ring_size - 1;
-                if constexpr(output_addr_offset != 0) {
+                if (output_addr_offset != 0) {
                     d.bank_base_address += last_output_addr_offset;
                 }
-                if constexpr(output_page_offset != 0) {
+                if (output_page_offset != 0) {
                     output_base_page_idx += last_output_page_offset;
                 }
             } else {
                 input_ring_idx--;
-                if constexpr(output_addr_offset != 0) {
+                if (output_addr_offset != 0) {
                     d.bank_base_address -= output_addr_offset;
                 }
-                if constexpr(output_page_offset != 0) {
+                if (output_page_offset != 0) {
                     output_base_page_idx -= output_page_offset;
                 }
             }
         } else {
             if (input_ring_idx == ring_size - 1) {
                 input_ring_idx = 0;
-                if constexpr(output_addr_offset != 0) {
+                if (output_addr_offset != 0) {
                     d.bank_base_address -= last_output_addr_offset;
                 }
-                if constexpr(output_page_offset != 0) {
+                if (output_page_offset != 0) {
                     output_base_page_idx -= last_output_page_offset;
                 }
             } else {
                 input_ring_idx++;
-                if constexpr(output_addr_offset != 0) {
+                if (output_addr_offset != 0) {
                     d.bank_base_address += output_addr_offset;
                 }
-                if constexpr(output_page_offset != 0) {
+                if (output_page_offset != 0) {
                     output_base_page_idx += output_page_offset;
                 }
             }