### Ticket
#11398 #11857

### Problem description
All gather and reduce scatter have fairly high dispatch time compared to other ops, sitting at ~20 us and 16 us each for Llama decode with tracing.

### Result
After optimization, llama decode dispatch latency for these ops has reduced to 8.5 us and 5 us respectively.

### What's changed:
AllGather:
- merged sender and receiver into the same kernel and uses rt arg to differentiate between modes
- all worker cores now launched at the same time with same ct args
- optimized set runtime args on host by getting the address of the runtime arg vector for all cores

ReduceScatter:
- all worker cores now launched at the same time with same ct args
- optimized set runtime args on host by getting the address of the runtime arg vector for all cores