@caixunshiren 2024-09-04
Hi @SeanNijjar feel free to review this branch now! It has now rebased on top of Ammar's changes on all gather matmul. Sorry for the delay -- was waiting for local tests to pass