@caixunshiren 2024-09-04
all post commit tests: https://github.com/tenstorrent/tt-metal/actions/runs/10693954147
t3k frequent tests: https://github.com/tenstorrent/tt-metal/actions/runs/10693975621
t3k nightly: https://github.com/tenstorrent/tt-metal/actions/runs/10693985724