@caixunshiren 2024-09-04
Found out a bug in sharded all gather matmul. Fixed and pushed again.

Here are the new set of tests:
all post commits: https://github.com/tenstorrent/tt-metal/actions/runs/10703346127
t3k frequent: https://github.com/tenstorrent/tt-metal/actions/runs/10703338829
t3k nightly: https://github.com/tenstorrent/tt-metal/actions/runs/10703323240