@caixunshiren 2024-09-04
rerun tests
all post commits: https://github.com/tenstorrent/tt-metal/actions/runs/10706751847
t3k frequent: https://github.com/tenstorrent/tt-metal/actions/runs/10706755866
t3k nightly: https://github.com/tenstorrent/tt-metal/actions/runs/10706763082